+++ mods/BrewLAN_RNG/RPS/mod_info.lua
@@ -1,0 +1,10 @@
+name = "Rock Paper Scissors"
+uid = "25D57D85-7D84-27HT-A502-RPS000000000"
+version = 0
+copyright = "© 2018 Sean Wheeldon"
+description = "Units are randomly assigned Rock, Paper, or Scissors. Damage is divided or multiplied by 4 based on standard rock paper scissors rules."
+author = "Balthazar"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false