+++ mods/BrewLAN_Units/BrewResearch/units/SAB0105/SAB0105_script.lua
@@ -1,0 +1,5 @@
+local ALandFactoryUnit = import('/lua/aeonunits.lua').ALandFactoryUnit
+
+SAB0305 = Class(ALandFactoryUnit) {}
+
+TypeClass = SAB0305